Rename number-theory to maths
