+++ include/dragon/maths/chinese-remainder-theorem.hpp
@@ -7,8 +7,8 @@
  * .
  * x*aN = 1 (mod mN)
  */
-#ifndef DRAGON_NUMBER_THEORY_CHINESE_REMAINDER_THEOREM_HPP
-#define DRAGON_NUMBER_THEORY_CHINESE_REMAINDER_THEOREM_HPP
+#ifndef DRAGON_MATHS_CHINESE_REMAINDER_THEOREM_HPP
+#define DRAGON_MATHS_CHINESE_REMAINDER_THEOREM_HPP
 #include "mod-inverse.hpp"
 #include <vector>
