+++ include/dragon/maths/extended-euclid-gcd.hpp
@@ -1,5 +1,5 @@
-#ifndef DRAGON_NUMBER_THEORY_EXTENDED_EUCLID_GCD_HPP
-#define DRAGON_NUMBER_THEORY_EXTENDED_EUCLID_GCD_HPP
+#ifndef DRAGON_MATHS_EXTENDED_EUCLID_GCD_HPP
+#define DRAGON_MATHS_EXTENDED_EUCLID_GCD_HPP
 #include <bits/stdc++.h>
 namespace dragon {
 /**
